bot/requirements.txt
discord.py==2.4.1
Flask==2.3.2
requests==2.31.0
